requirements.txt
Pillow==8.3.1
pystray==0.17.3
python-xlib==0.31
six==1.16.0
